+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Improvements
 
 - Threading issues in internal dependency container (#5225)
+- Add auto-start for session tracker when SDK is started after didBecomeActive (#5121)
 
 ## 8.51.0
 
@@ -18,7 +19,6 @@
 ### Fixes
 
 - Reporting unhandled C++ exception message (#5190)
-- Add auto-start for session tracker when SDK is started after didBecomeActive (#5121)
 - Improved internal multi-threading of session replay to fix thread inversion warning and reduce chance of queue starvation (#5018)
 
 ### Improvements